requirements.txt
Django==2.2.7
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.11.0
djangorestframework-jwt-custom-user==1.11.1
djangorestframework-simplejwt==4.4.0
